requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
celery==5.4.0
redis==5.1.0
scikit-learn==1.5.2
datasets==3.0.1
transformers==4.45.1
weasyprint==62.3
nbformat==5.10.4
jinja2==3.1.4
pandas==2.2.3
numpy==2.1.1
python-dotenv==1.0.1
httpx==0.27.2
torch==2.4.1
supabase==2.11.0
